backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5
markdown==3.5.2
drf-yasg==1.21.7
gunicorn==21.2.0
